requirements.txt
pyvts
sherpa-onnx
sounddevice
pyyaml
loguru
requests
tqdm
onnxruntime==1.20.0
webrtcvad-wheels
PyQt6
qasync
opencv-python-headless